Move ssim, tesseract testing to unit tests
