+++ .github/workflows/ci.yml
@@ -12,6 +12,8 @@ jobs:
       run:
         shell: bash -l {0}
     steps:
+      - name: Dependencies
+        run: sudo apt update && sudo apt install ffmpeg libsm6 libxext6 tesseract-ocr tesseract-ocr-eng libtesseract-dev libleptonica-dev pkg-config -y
       - uses: actions/checkout@v4
       - uses: conda-incubator/setup-miniconda@v3
         with:
@@ -20,6 +22,7 @@ jobs:
           python-version: 3.11
           auto-activate-base: false
       - name: Test with pytest
+        # export AICOACH_STUDENT__NAME="zatic" AICOACH_TESSDATA_DIR="/usr/share/tesseract-ocr/4.00/tessdata/" AICOACH_MONGO_DSN="mongodb://localhost:27017" AICOACH_ASSISTANT_ID="XXX" AICOACH_OPENAI_API_KEY="YYY" AICOACH_OPENAI_ORG_ID="ZZZ" AICOACH_OBS_INTEGRATION="False" AICOACH_AUDIOMODE="text" && pytest tests/unit --doctest-modules --junitxml=tests/test-results.xml --cov=. --cov-report=xml --cov-report=html
         run: |
           pytest tests/unit --doctest-modules --junitxml=tests/test-results.xml --cov=. --cov-report=xml --cov-report=html
         env:
@@ -27,6 +30,10 @@ jobs:
           AICOACH_ASSISTANT_ID: "XXX"
           AICOACH_OPENAI_API_KEY: "YYY"
           AICOACH_OPENAI_ORG_ID: "ZZZ"
+          AICOACH_TESSDATA_DIR: /usr/share/tesseract-ocr/4.00/tessdata/
+          AICOACH_OBS_INTEGRATION: "False"
+          AICOACH_AUDIOMODE: "text"
+          AICOACH_STUDENT__NAME: "zatic"
       - name: Test Report
         uses: dorny/test-reporter@v1
         if: success() || failure()
