+++ external/fast_ssim/libssim.so
