+++ obs_tools/__init__.py
@@ -1,7 +1,6 @@
 from config import AudioMode, config
 
 if config.obs_integration:
-    from .mapstats import get_map_stats
     from .parse_map_loading_screen import LoadingScreenScanner as GameStartedScanner
 else:
     from .sc2client import ClientAPIScanner as GameStartedScanner
